Adding Arduino class as well as the initial program flow
This adds a new ArduinoSerial class that is a very simple wrapper around
the serial port for the Arduino. The biggest thing it does is allow us
to use a list of commands for easy sending of commands. While we could
very easily just send a string (ie . "N198EN...") the problem is that
the E6000 requires a pause between commands, which this handles for us.